+++ shared/multisig.py
@@ -757,7 +757,7 @@ class MultisigWallet:
             # HACK but there is no difference extended_keys - just bech32 hrp
             assert chain.ctype == "XTN"
         else:
-            assert chain.ctype == expect_chain      # 'wrong chain'
+            assert chain.ctype == expect_chain, 'wrong chain'
 
         depth = node.depth()