+++ releases/ChangeLog.md
@@ -6,6 +6,8 @@
   [BIP-0380](https://github.com/bitcoin/bips/blob/master/bip-0380.mediawiki#key-expressions)
   was `(m=0F056943)/m/48'/1'/0'/2'/0/0` now `[0F056943/48'/1'/0'/2'/0/0]`
 - Enhancement: Address explorer UX cosmetics, now with arrows and dots.
+- Enhancement: Linked settings (multisig, trick pins, backup password, hsm users and utxo cache)
+  separation for new main secret.
 - Rename `Unchained Capital` to `Unchained`
 - Bugfix: Correct `scriptPubkey` parsing for segwit v1-v16
 - Bugfix: Do not infer segwit just by availability of `PSBT_IN_WITNESS_UTXO` in PSBT.
@@ -14,6 +16,7 @@
 - Bugfix: Empty number during BIP-39 passphrase entry could cause crash.
 - Bugfix: Signing with BIP39 Passphrase showed master fingerprint as integer. Fixed to show hex.
 - Bugfix: Fixed inability to generate paper wallet without secrets
+- Bugfix: Activating trick pin duress wallet copied multisig settings from main wallet
 - Bugfix: SD2FA setting is cleared when seed is wiped after failed login due to policy SD2FA enforce.
   Prevents infinite seed wipe loop when restoring backup after 2FA MicroSD lost or damaged.
   SD2FA is not backed up and also not restored from older backups. If SD2FA is set up,